requirements.txt
streamlit
pandas
selenium
beautifulsoup4
scikit-learn
nltk
google-generativeai